fix(#4096): move the call-dispatch tail out of the god-file; record the measurement
Two things, both consequences of the first commit:

1. The LOC ratchet correctly refused +15 lines in `call-tail-dispatch.ts`. The
   right answer was not an allowance: the graceful `ref.null.extern` fallback —
   the point where "no arm recognised this call" is silently rendered as the
   VALUE `undefined` — moves into `stored-member-closure-call.ts` together with
   the arm that narrows it, so the silence and its narrowings are read at once.
   The god-file gets a single `return compileCallDispatchTail(...)` and shrinks
   by 12 lines.

2. #4096 written up as a RESOLUTION, including the two ways the original
   scoping was wrong, because both cost real time:

   - The site was NOT unfindable by marker. It is a plain single-statement
     `fctx.body.push({ op: "ref.null.extern" })`, i.e. inside the class the
     earlier 463-site sweep claimed to have covered and found nothing in. A
     sweep that reports "absent" everywhere is indistinguishable from one that
     never ran; it needed a positive control.
   - The defect is NOT "a member absent from a closed static type". The member
     is present and correct — the read yields the right function and two
     successive assignments last-write-win. Only the direct call form was
     broken, and for a plain user function as much as for a String.prototype
     transfer. The trigger is any assignment-STORED function member, which is
     much more ordinary JavaScript than the issue's framing.

Measured, scoped standalone A/B over the 144 shape-matched
`built-ins/String/prototype` files (runTest262File status, NOT the CI path):
**+21 fail->pass, 0 pass->fail, 0 other status changes.** Equivalence shard 1/8
reports no new regressions. The "34 shape-matched files is the residue, not a
flip prediction" sentence stays true and is now also the wrong denominator —
noted as such rather than quietly reused.

Residuals filed, not folded in: #4113 (the same stored member is invisible to
hasOwnProperty / `in` / for-in / computed read while the static read sees it —
a different subsystem, #4086/#4010 territory, and NOT provably non-displacing
the way this arm is), and a RegExp receiver whose member READ is already wrong
so there is no correct value for any call arm to invoke.
